requirements.txt
phidata
duckduckgo-search
googlesearch-python
newspaper4k
requests
lxml_html_clean
pypdf
sqlalchemy
sentence-transformers
lancedb
pyarrow
groq
ollama
trafilatura
beautifulsoup4
colorama
crawl4ai
asyncio
python-dotenv